// src/ad-screens/types/ad-screen.types.ts
export enum LayoutType {
  landscape_full = 'Landscape-Full',
  landscape_80_20 = 'Landscape-80/20',
  landscape_20_80 = 'Landscape-20/80',
  landscape_split = 'Landscape-Split',
  portrait_full = 'Portrait-Full',
  portrait_80_20 = 'Portrait-80/20',
  portrait_20_80 = 'Portrait-20/80',
  portrait_split = 'Portrait-Split',
}

export enum Status {
  offline = 0,
  online = 1,
}